refactor: atualizar Exceptions
- Atualiza ApplicationException, AuthenticationServiceException
- Melhora BadRequestException, ConflictException, NotFoundException
- Padroniza InternalException, RequestTimeoutException
- Atualiza UnauthorizedException e ValidationException

+++ src/Riber.Application/Exceptions/AuthenticationServiceException.cs
@@ -3,4 +3,4 @@ using System.Net;
 namespace Riber.Application.Exceptions;
 
 public sealed class AuthenticationServiceException(string message) 
-    : ApplicationException(message, (int)HttpStatusCode.InternalServerError);
+    : ApplicationException(message, HttpStatusCode.InternalServerError);
