+++ src/Riber.Application/Exceptions/ConflictException.cs
@@ -3,4 +3,4 @@ using System.Net;
 namespace Riber.Application.Exceptions;
 
 public sealed class ConflictException(string message) 
-    : ApplicationException(message, (int)HttpStatusCode.Conflict);
+    : ApplicationException(message, HttpStatusCode.Conflict);
