+++ src/Riber.Application/Exceptions/NotFoundException.cs
@@ -3,4 +3,4 @@ using System.Net;
 namespace Riber.Application.Exceptions;
 
 public sealed class NotFoundException(string message)
-    : ApplicationException(message, (int)HttpStatusCode.NotFound);
+    : ApplicationException(message, HttpStatusCode.NotFound);
