+++ src/Riber.Application/Exceptions/RequestTimeoutException.cs
@@ -3,4 +3,4 @@
 namespace Riber.Application.Exceptions;
 
 public sealed class RequestTimeoutException(string requestName, TimeSpan elapsedTime)
-    : ApplicationException($"Request '{requestName}' timed out after {elapsedTime.TotalSeconds:F1} seconds", (int)HttpStatusCode.RequestTimeout);
+    : ApplicationException($"Request '{requestName}' timed out after {elapsedTime.TotalSeconds:F1} seconds", HttpStatusCode.RequestTimeout);
