+++ src/Riber.Application/Exceptions/UnauthorizedException.cs
@@ -3,4 +3,4 @@ using System.Net;
 namespace Riber.Application.Exceptions;
 
 public sealed class UnauthorizedException(string message)
-    : ApplicationException(message, (int)HttpStatusCode.Unauthorized);
+    : ApplicationException(message, HttpStatusCode.Unauthorized);
